+++ game_patch/rf/player.h
@@ -33,8 +33,9 @@ namespace rf
     };
     static_assert(sizeof(Camera) == 0xC);
 
-    static auto& CameraSetFirstPerson = AddrAsRef<void(Camera *camera)>(0x0040DDF0);
-    static auto& CameraSetFreelook = AddrAsRef<void(Camera *camera)>(0x0040DCF0);
+    static auto& CameraSetFirstPerson = AddrAsRef<bool(Camera *camera)>(0x0040DDF0);
+    static auto& CameraSetFreelook = AddrAsRef<bool(Camera *camera)>(0x0040DCF0);
+    static auto& CameraSetFixed = AddrAsRef<bool(Camera *camera)>(0x0040DF70);
     static auto& CameraGetPos = AddrAsRef<void(rf::Vector3* camera_pos, rf::Camera *camera)>(0x0040D760);
     static auto& CameraGetOrient = AddrAsRef<void(rf::Matrix3* camera_orient, rf::Camera *camera)>(0x0040D780);